# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cromos API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Registrar nuevo usuario",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario registrado con éxito"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Autenticar usuario",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario autenticado correctamente"
                    },
                    "401": {
                        "description": "Credenciales inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Obtener datos del usuario autenticado",
                "operationId": "e44b931999dcfc054d7a30b300537f43",
                "responses": {
                    "200": {
                        "description": "Datos del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Juan Pérez"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "juan@example.com"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2024-05-15T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/collections": {
            "get": {
                "tags": [
                    "Colecciones"
                ],
                "summary": "Obtener todas las colecciones",
                "description": "Retorna una lista de todas las colecciones disponibles.",
                "operationId": "1c56670af32b2e0bb7c56f75d42e7789",
                "responses": {
                    "200": {
                        "description": "Listado de colecciones obtenido con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "collections": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Coleccion"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/subscribedCollections": {
            "get": {
                "tags": [
                    "Colecciones"
                ],
                "summary": "Obtener todas las colecciones del usuario",
                "description": "Retorna una lista de todas las colecciones a las que el usuario se ha suscrito.",
                "operationId": "74445568c789650dda4f840d219296ac",
                "responses": {
                    "200": {
                        "description": "Listado de colecciones obtenido con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "collections": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Coleccion"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cards/{collection}": {
            "get": {
                "tags": [
                    "Colecciones"
                ],
                "summary": "Obtener todos los cromos de una colección",
                "description": "Retorna una lista de todos los cromos de una colección.",
                "operationId": "2ec15b354aaf72060f124317c49c68f3",
                "parameters": [
                    {
                        "name": "collection",
                        "in": "path",
                        "description": "ID de la colección",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de cromos obtenido con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cards": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Coleccion"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Coleccion": {
                "title": "Colección",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Colección de Verano"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Endpoints para registrar, autenticar y manejar sesión de usuarios"
        },
        {
            "name": "Colecciones",
            "description": "Colecciones"
        }
    ]
}